fix: all reuseable checks added
