+++ dashboard_dependencies_ymls/datafiles/django-splash_dependencies_health.yaml
@@ -1,4 +1,70 @@
 TIMESTAMP: 2023-12-01
+dependencies:
+    count: 98
+    github:
+        count: 0
+        list: '[]'
+    js:
+        count: 0
+        list: ''
+    js.all:
+        count: 0
+        list: ''
+    js.dev:
+        count: 0
+        list: ''
+    pypi:
+        count: 85
+        list: '["chardet==5.2.0", "lxml==4.9.3", "py==1.11.0", "docutils==0.20.1",
+            "readme-renderer==42.0", "django==3.2.22", "certifi==2023.7.22", "edx-i18n-tools==1.3.0",
+            "sqlparse==0.4.4", "urllib3==2.0.6", "nh3==0.2.14", "caniusepython3==7.3.0",
+            "edx-lint==5.3.4", "click==8.1.7", "diff-cover==7.7.0", "django-crum==0.7.9",
+            "more-itertools==10.1.0", "pkginfo==1.9.6", "secretstorage==3.3.3", "isort==5.12.0",
+            "pytz==2023.3.post1", "django-waffle==4.0.0", "newrelic==9.1.0", "markupsafe==2.1.3",
+            "cryptography==41.0.4", "jeepney==0.8.0", "jinja2==3.1.2", "lazy-object-proxy==1.9.0",
+            "platformdirs==3.11.0", "code-annotations==1.5.0", "idna==3.4", "edx-django-utils==5.7.0",
+            "importlib-resources==6.1.0", "mccabe==0.7.0", "requests-toolbelt==1.0.0",
+            "jaraco-classes==3.3.0", "path==16.7.1", "djangorestframework==3.14.0",
+            "pydocstyle==6.3.0", "wrapt==1.15.0", "pylint-django==2.5.3", "backports-functools-lru-cache==1.6.6",
+            "stevedore==5.1.0", "pycparser==2.21", "rfc3986==2.0.0", "tomli==2.0.1",
+            "twine==4.0.2", "typing-extensions==4.8.0", "snowballstemmer==2.2.0",
+            "keyring==24.2.0", "astroid==2.15.8", "pygments==2.16.1", "requests==2.31.0",
+            "mdurl==0.1.2", "charset-normalizer==3.3.0", "asgiref==3.7.2", "pbr==5.11.1",
+            "pylint-plugin-utils==0.8.2", "packaging==23.2", "pycodestyle==2.11.0",
+            "virtualenv==20.24.5", "django-config-models==2.5.1", "distlib==0.3.7",
+            "filelock==3.12.4", "tomlkit==0.12.1", "click-log==0.4.0", "pluggy==1.3.0",
+            "text-unidecode==1.3", "zipp==3.17.0", "pylint-celery==0.3", "six==1.16.0",
+            "rich==13.6.0", "pyyaml==6.0.1", "pylint==2.17.7", "dill==0.3.7", "psutil==5.9.5",
+            "markdown-it-py==3.0.0", "polib==1.2.0", "tox==3.28.0", "importlib-metadata==6.8.0",
+            "tox-battery==0.6.2", "cffi==1.16.0", "wheel==0.41.2", "pynacl==1.5.0",
+            "python-slugify==8.0.1"]'
+    pypi_all:
+        count: 98
+        list: '["setuptools==68.2.2", "chardet==5.2.0", "lxml==4.9.3", "py==1.11.0",
+            "docutils==0.20.1", "readme-renderer==42.0", "django==3.2.22", "certifi==2023.7.22",
+            "edx-i18n-tools==1.3.0", "sqlparse==0.4.4", "urllib3==2.0.6", "pytest-cov==4.1.0",
+            "nh3==0.2.14", "caniusepython3==7.3.0", "edx-lint==5.3.4", "click==8.1.7",
+            "diff-cover==7.7.0", "django-crum==0.7.9", "more-itertools==10.1.0", "pkginfo==1.9.6",
+            "codecov==2.1.13", "coverage[toml]==7.3.2", "secretstorage==3.3.3", "exceptiongroup==1.1.3",
+            "isort==5.12.0", "pytz==2023.3.post1", "django-waffle==4.0.0", "pytest-django==4.5.2",
+            "newrelic==9.1.0", "markupsafe==2.1.3", "pyproject-hooks==1.0.0", "cryptography==41.0.4",
+            "jeepney==0.8.0", "jinja2==3.1.2", "lazy-object-proxy==1.9.0", "coverage==7.3.2",
+            "platformdirs==3.11.0", "code-annotations==1.5.0", "idna==3.4", "edx-django-utils==5.7.0",
+            "importlib-resources==6.1.0", "mccabe==0.7.0", "build==1.0.3", "requests-toolbelt==1.0.0",
+            "jaraco-classes==3.3.0", "path==16.7.1", "djangorestframework==3.14.0",
+            "pydocstyle==6.3.0", "wrapt==1.15.0", "pylint-django==2.5.3", "backports-functools-lru-cache==1.6.6",
+            "stevedore==5.1.0", "pycparser==2.21", "rfc3986==2.0.0", "tomli==2.0.1",
+            "twine==4.0.2", "typing-extensions==4.8.0", "snowballstemmer==2.2.0",
+            "keyring==24.2.0", "astroid==2.15.8", "pygments==2.16.1", "requests==2.31.0",
+            "mdurl==0.1.2", "charset-normalizer==3.3.0", "asgiref==3.7.2", "pip==23.2.1",
+            "pbr==5.11.1", "pylint-plugin-utils==0.8.2", "packaging==23.2", "pycodestyle==2.11.0",
+            "virtualenv==20.24.5", "iniconfig==2.0.0", "pytest==7.4.2", "django-config-models==2.5.1",
+            "distlib==0.3.7", "filelock==3.12.4", "tomlkit==0.12.1", "click-log==0.4.0",
+            "pluggy==1.3.0", "text-unidecode==1.3", "zipp==3.17.0", "pylint-celery==0.3",
+            "six==1.16.0", "rich==13.6.0", "pyyaml==6.0.1", "pylint==2.17.7", "dill==0.3.7",
+            "pip-tools==7.3.0", "psutil==5.9.5", "markdown-it-py==3.0.0", "polib==1.2.0",
+            "tox==3.28.0", "importlib-metadata==6.8.0", "tox-battery==0.6.2", "cffi==1.16.0",
+            "wheel==0.41.2", "pynacl==1.5.0", "python-slugify==8.0.1"]'
 django:
     '4.0': 1.2.0
     '4.1': null
@@ -7,18 +73,18 @@ django:
 django_packages:
     django_32:
         count: 10
-        list: '["django-crum", "edx-i18n-tools", "django-config-models", "pytest-django",
-            "newrelic", "django-waffle", "edx-django-utils", "code-annotations", "pylint-django",
+        list: '["edx-i18n-tools", "pytest-django", "newrelic", "django-waffle", "pylint-django",
+            "django-crum", "django-config-models", "code-annotations", "edx-django-utils",
             "djangorestframework"]'
     total:
         count: 10
-        list: '["django-crum", "edx-i18n-tools", "django-config-models", "pytest-django",
-            "newrelic", "django-waffle", "edx-django-utils", "code-annotations", "pylint-django",
+        list: '["edx-i18n-tools", "pytest-django", "newrelic", "django-waffle", "pylint-django",
+            "django-crum", "django-config-models", "code-annotations", "edx-django-utils",
             "djangorestframework"]'
     upgraded:
         count: 10
-        list: '["django-crum", "edx-i18n-tools", "django-config-models", "pytest-django",
-            "newrelic", "django-waffle", "edx-django-utils", "code-annotations", "pylint-django",
+        list: '["edx-i18n-tools", "pytest-django", "newrelic", "django-waffle", "pylint-django",
+            "django-crum", "django-config-models", "code-annotations", "edx-django-utils",
             "djangorestframework"]'
 github:
     branch_count: 47
