+++ requirements/test.in
@@ -4,3 +4,4 @@
 -r base.txt               # Core dependencies for this package
 responses
 pytest-cov
+toml
